+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 # Changelog
 
+### 2.0.0 (wip)
+
+- Maze machine.
+
 ### 1.6.0 (wip)
 
-...
+- Use ChaCha20 instead of AES for encryption.
+- Use Nuitka instead of Cython for compilation.
 
 ### 1.5.2 (2026-06-30)